Fixed conflicts and run unit testing.
