+++ src/DwitTech.AccountService.Core/Services/ActivationService.cs
@@ -1,12 +1,7 @@
 ﻿using DwitTech.AccountService.Core.Interfaces;
-using DwitTech.AccountService.Core.Models;
 using DwitTech.AccountService.Core.Utilities;
 using DwitTech.AccountService.Data.Repository;
 using Microsoft.Extensions.Configuration;
-using Newtonsoft.Json;
-using RestSharp;
-using System.Security.Cryptography.X509Certificates;
-using System.Text.Json.Serialization;
 
 namespace DwitTech.AccountService.Core.Services
 {
@@ -14,13 +9,18 @@ namespace DwitTech.AccountService.Core.Services
     {
         private readonly IConfiguration _configuration;
         private readonly IUserRepository _userRepository;
+        private readonly IEmailService _emailService;
 
 
-        public ActivationService(IConfiguration configuration, IUserRepository userRepository)
+        public ActivationService(IConfiguration configuration, IUserRepository userRepository, IEmailService emailService)
         {
             _configuration = configuration;
             _userRepository = userRepository;
+            _emailService = emailService;
+        }
 
+        public ActivationService(IConfigurationRoot configuration, IUserRepository @object)
+        {
         }
 
         private static string GetActivationCode()
@@ -59,37 +59,19 @@ namespace DwitTech.AccountService.Core.Services
             return templateText.ToString();
         }
 
-        private async Task<bool> SendMailAsync(string fromEmail, string toEmail, string subject, string body, string cc, string bcc)
-        {
-            var emailObject = new Email { FromEmail = fromEmail, ToEmail = toEmail, Subject = subject, Body = body };
-
-            var url = "https://jsonplaceholder.typicode.com/posts";
-
-
-            var client = new RestClient(url);
-            var request = new RestRequest(url, Method.Post); 
-            request.AddHeader("Content-Type", "application/json");
-            var message = JsonConvert.SerializeObject(emailObject);
-            request.AddBody(message, "application/json");
-            RestResponse response = await client.ExecuteAsync(request);
-
-            if (response is null)
-                return false;
-            
-            return true;
-
-        }
-
+        
         public async Task<bool> SendActivationEmail(string userId, string fromEmail, string toEmail, string templateName, string RecipientName,
-            string subject = "Account Activation", string cc = "", string bcc = "")
+            string cc = "", string bcc = "")
         {
-
-            var activationUrl = await GetActivationUrl(userId);
+            const string Subject = "Account Activation";
+           
+           var activationUrl = await GetActivationUrl(userId);
             string templateText = GetTemplate(templateName);
             templateText = templateText.Replace("{{name}}", RecipientName);
             templateText = templateText.Replace("{{activationUrl}}", activationUrl);
             string body = templateText;
-            var response = await SendMailAsync(fromEmail, toEmail, subject, body, cc, bcc);
+            var response = await _emailService.SendMailAsync(userId, fromEmail, toEmail, Subject, body, cc, bcc);
+         
 
             return response;
         }
